better tab completion
When this option is turned on the using tab for command-line completion will
also display all the possible matches on your status line. You can then tab
through those matches and select the one you are looking for. It can match on
just about anything you can type as a command. For example type “:col” and
hitting tab will display colder and colorscheme. Typing “:colorscheme a” and
hitting tab will list all the colorschemes that start with a. This gets really
handy when typing out long directory paths as it will list files and
directories of the current path you are typing (and save you from having to
type the entire thing out yourself).